useState hook limitation preventing multiple state updates
Previously, our custom implementation of the useState hook only allowed a single update to a state variable on each render cycle, even if multiple
updates were triggered in quick succession. This was due to the fact that setState was always bound to the original StateMachine, even though updating
the state triggers a fork into a new instance.

The fix was to support "StateFactories" when forking, allowing the "fork" method to not only copy existing state, but to also recreate it given the
receiving StateMachine. That way, when setState triggers a fork, the fork actually contains a factory that recreates the "setState" function, but this
time bound to the new StateMachine.

This behavior was covered with a test.